Move Routes to a component
