+++ dashboard/src/components/Routes/index.tsx
@@ -1,0 +1,3 @@
+import Routes from "./Routes";
+
+export default Routes;
