+++ dashboard/src/containers/RoutesContainer/index.ts
@@ -1,0 +1,3 @@
+import RoutesContainer from "./RoutesContainer";
+
+export default RoutesContainer;
